[SYCL] Split image accessor code into a separate include
... that isn't included from `<sycl/detail/core.hpp>`
