+++ sycl/test-e2e/Basic/image/image_accessor_readwrite.cpp
@@ -16,6 +16,7 @@
 
 #include <cassert>
 #include <iomanip>
+#include <sycl/accessor_image.hpp>
 #include <sycl/detail/core.hpp>
 #if DEBUG_OUTPUT
 #include <iostream>
