+++ sycl/test-e2e/Regression/image_access.cpp
@@ -15,6 +15,7 @@
 
 #include <iostream>
 
+#include <sycl/accessor_image.hpp>
 #include <sycl/detail/core.hpp>
 
 int main() {
