+++ sycl/test/basic_tests/image_accessor_types.cpp
@@ -13,10 +13,10 @@ int main() {
   myQueue.submit([&](handler &cgh) {
     accessor<float, 2, access::mode::read, access::target::image> NotValidType1(
         srcImage, cgh);
-    // expected-error@sycl/accessor.hpp:* {{The data type of an image accessor must be only cl_int4, cl_uint4, cl_float4 or cl_half4}}
+    // expected-error@sycl/accessor_image.hpp:* {{The data type of an image accessor must be only cl_int4, cl_uint4, cl_float4 or cl_half4}}
     accessor<int2, 2, access::mode::read, access::target::image> NotValidType2(
         srcImage, cgh);
-    // expected-error@sycl/accessor.hpp:* {{The data type of an image accessor must be only cl_int4, cl_uint4, cl_float4 or cl_half4}}
+    // expected-error@sycl/accessor_image.hpp:* {{The data type of an image accessor must be only cl_int4, cl_uint4, cl_float4 or cl_half4}}
     accessor<float4, 2, access::mode::read, access::target::image>
         ValidSYCLFloat(srcImage, cgh);
     accessor<int4, 2, access::mode::read, access::target::image> ValidSYCLInt(
