+++ sycl/test/include_deps/sycl_accessor.hpp.cpp
@@ -113,9 +113,10 @@
 // CHECK-NEXT: detail/accessor_iterator.hpp
 // CHECK-NEXT: detail/generic_type_traits.hpp
 // CHECK-NEXT: detail/handler_proxy.hpp
-// CHECK-NEXT: detail/image_accessor_util.hpp
 // CHECK-NEXT: device.hpp
 // CHECK-NEXT: kernel_bundle_enums.hpp
+// CHECK-NEXT: pointers.hpp
+// CHECK-NEXT: properties/accessor_properties.hpp
 // CHECK-NEXT: image.hpp
 // CHECK-NEXT: detail/backend_traits.hpp
 // CHECK-NEXT: sampler.hpp
@@ -125,8 +126,5 @@
 // CHECK-NEXT: detail/memcpy.hpp
 // CHECK-NEXT: detail/vector_convert.hpp
 // CHECK-NEXT: swizzles.def
-// CHECK-NEXT: pointers.hpp
-// CHECK-NEXT: properties/accessor_properties.hpp
 // CHECK-NEXT: properties/buffer_properties.hpp
-// CHECK-NEXT: detail/image_ocl_types.hpp
 // CHECK-EMPTY:
